[core] Расширение интерфейсов [import-mimio] Импорт групп, изображений, прямоугольников и эллипсов

+++ core/board-common/src/shape.cpp
@@ -81,7 +81,7 @@ void Shape::setInnerSize(const QSizeF &size)
   emit scaleyChanged();
 }
 
-void Shape::setThickness(int thickness)
+void Shape::setThickness(float thickness)
 {
   _thickness = thickness;
   emit thicknessChanged();
